+++ packages/flame/lib/src/particles/circle_particle.dart
@@ -2,20 +2,19 @@ import 'dart:ui';
 
 import 'particle.dart';
 
-/// Plain circle with no other behaviors
-/// Consider composing with other [Particle]
-/// to achieve needed effects
+/// Plain circle with no other behaviors.
+///
+/// Consider composing with other [Particle] to achieve needed effects.
 class CircleParticle extends Particle {
   final Paint paint;
+
   final double radius;
 
   CircleParticle({
     required this.paint,
     this.radius = 10.0,
     double? lifespan,
-  }) : super(
-          lifespan: lifespan,
-        );
+  }) : super(lifespan: lifespan);
 
   @override
   void render(Canvas c) {